git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@224 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/TEST/ContinuousWaveletTransform_test.C
@@ -259,7 +259,7 @@ CHECK((void init(double scale, double spacing, unsigned int MZ)))
   Test cwt;
   double scale = 0.2;
   double spacing = 2.3;
-  double mz = 2;
+  unsigned int mz = 2;
   cwt.init(scale,spacing,mz);
   
   TEST_EQUAL(cwt.getMzDim(), mz)
